requirements.txt
numpy
ripser
step
tqdm
scikit-learn
matplotlib
persim
opencv-python-headless
Pillow
